+++ dart/test/sentry_event_test.dart
@@ -129,7 +129,6 @@ void main() {
           'platform': isWeb ? 'javascript' : 'other',
           'event_id': '00000000000000000000000000000000',
           'timestamp': '2019-01-01T00:00:00.000Z',
-          'sdk': {'version': sdkVersion, 'name': sdkName},
           'message': {
             'formatted': 'test-message 1 2',
             'message': 'test-message %d %d',
@@ -257,6 +256,34 @@ void main() {
           'handled': true,
         },
       );
+    });
+
+    test('should not serialize null or empty fields', () {
+      final event = SentryEvent(
+        message: Message(null),
+        modules: {},
+        exception: SentryException(type: null, value: null),
+        stackTrace: SentryStackTrace(frames: []),
+        tags: {},
+        extra: {},
+        contexts: Contexts(),
+        fingerprint: [],
+        breadcrumbs: [Breadcrumb()],
+        request: Request(),
+        debugMeta: DebugMeta(images: []),
+      );
+      final eventMap = event.toJson();
+
+      expect(eventMap['message'], isNull);
+      expect(eventMap['modules'], isNull);
+      expect(eventMap['exception'], isNull);
+      expect(eventMap['stacktrace'], isNull);
+      expect(eventMap['tags'], isNull);
+      expect(eventMap['extra'], isNull);
+      expect(eventMap['contexts'], isNull);
+      expect(eventMap['fingerprint'], isNull);
+      expect(eventMap['request'], isNull);
+      expect(eventMap['debug_meta'], isNull);
     });
   });
 }